agents copy/requirements.txt
# Core validation dependencies
jsonschema>=4.19
pytest>=7.4

# Code quality tools
ruff>=0.5
black>=24.0

# Optional dependencies (for full functionality)
pandas>=2.0
numpy>=1.24
geopandas>=0.13
shapely>=2.0
pandapipes>=0.8
pandapower>=2.14
pyyaml>=6.0
openpyxl>=3.1
scikit-learn>=1.3
lightgbm>=4.0
pyarrow>=12.0  # For parquet files

# Development dependencies
pytest-cov>=4.0 